+++ lib/src/day_view/day_view.dart
@@ -226,6 +226,9 @@ class DayView<T extends Object?> extends StatefulWidget {
   /// This field will be used to set end hour for day view
   final int endHour;
 
+  /// Flag to keep scrollOffset of pages on page change
+  final bool keepScrollOffset;
+
   /// Main widget for day view.
   const DayView({
     Key? key,
@@ -275,6 +278,7 @@ class DayView<T extends Object?> extends StatefulWidget {
     this.emulateVerticalOffsetBy = 0,
     this.onEventDoubleTap,
     this.endHour = Constants.hoursADay,
+    this.keepScrollOffset = false,
   })  : assert(!(onHeaderTitleTap != null && dayTitleBuilder != null),
             "can't use [onHeaderTitleTap] & [dayTitleBuilder] simultaneously"),
         assert(timeLineOffset >= 0,
@@ -350,7 +354,7 @@ class DayViewState<T extends Object?> extends State<DayView<T>> {
   @override
   void initState() {
     super.initState();
-    _lastScrollOffset = widget.scrollOffset;
+    _lastScrollOffset = widget.scrollOffset ?? 0.0;
 
     _reloadCallback = _reload;
     _setDateRange();
@@ -486,8 +490,6 @@ class DayViewState<T extends Object?> extends State<DayView<T>> {
                             minuteSlotSize: widget.minuteSlotSize,
                             scrollNotifier: _scrollConfiguration,
                             fullDayEventBuilder: _fullDayEventBuilder,
-                            scrollOffset: _lastScrollOffset,
-                          scrollListener: _scrollPageListener,
                             showHalfHours: widget.showHalfHours,
                             showQuarterHours: widget.showQuarterHours,
                             halfHourIndicatorSettings:
@@ -498,6 +500,10 @@ class DayViewState<T extends Object?> extends State<DayView<T>> {
                                 _quarterHourIndicatorSettings,
                             emulateVerticalOffsetBy:
                                 widget.emulateVerticalOffsetBy,
+                            lastScrollOffset: _lastScrollOffset,
+                            dayViewScrollController: _scrollController,
+                            scrollListener: _scrollPageListener,
+                            keepScrollOffset: widget.keepScrollOffset,
                           ),
                         );
                       },
@@ -745,7 +751,9 @@ class DayViewState<T extends Object?> extends State<DayView<T>> {
         _currentIndex = index;
       });
     }
-    animateToDuration(widget.startDuration);
+    if (!widget.keepScrollOffset) {
+      animateToDuration(widget.startDuration);
+    }
     widget.onPageChange?.call(_currentDate, _currentIndex);
   }